+++ spotbugs/src/main/java/edu/umd/cs/findbugs/DetectorFactory.java
@@ -122,6 +122,10 @@ public class DetectorFactory {
 
     private final boolean defEnabled;
 
+    /**
+     * @deprecated This attribute is not used actively, and could be removed in future release
+     */
+    @Deprecated
     private final String speed;
 
     private final String reports;
@@ -320,6 +324,7 @@ public class DetectorFactory {
 
     /**
      * Get the speed of the Detector produced by this factory.
+     * @deprecated This attribute is not used actively, and could be removed in future release
      */
     @Deprecated
     public String getSpeed() {