+++ spotbugs/src/main/java/edu/umd/cs/findbugs/filter/SourceMatcher.java
@@ -23,14 +23,16 @@ package edu.umd.cs.findbugs.filter;
 
 import java.io.IOException;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import edu.umd.cs.findbugs.BugInstance;
 import edu.umd.cs.findbugs.ClassAnnotation;
-import edu.umd.cs.findbugs.SystemProperties;
 import edu.umd.cs.findbugs.xml.XMLAttributeList;
 import edu.umd.cs.findbugs.xml.XMLOutput;
 
 public class SourceMatcher implements Matcher {
-    private static final boolean DEBUG = SystemProperties.getBoolean("filter.debug");
+    private static final Logger LOG = LoggerFactory.getLogger(SourceMatcher.class);
 
     private final NameMatch fileName;
 
@@ -62,10 +64,9 @@ public class SourceMatcher implements Matcher {
                 result = fileName.match(bugFileName);
             }
         }
-        
-        if (DEBUG) {
-            System.out.println("Matching " + bugFileName + " with " + fileName + ", result = " + result);
-        }
+
+        result = fileName.match(bugFileName);
+        LOG.debug("Matching {} with {}, result = {}", bugFileName, fileName, result);
         return result;
     }