+++ CHANGELOG.md
@@ -20,6 +20,16 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
 * Start migrating STDOUT/STDERR usage to a logging framework
 * Improvements and bug-fixes for fancy-hist.xsl
 
+### Deprecated
+
+* SQL files
+* JNLP files
+* `speed` attribute of `Detector` element in `findbugs.xml`
+
+### Fixed
+
+* Fixed bug priority calculation logic in FindNonShortCircuit#reportBug
+
 ## 3.1.12 - 2019-02-28
 
 ### Added